Ensure Sentinel#master_for propagates connection errors (#1697)
* Ensure Sentinel#master_for propagates connection errors

master_for will first ask sentinel for the master and then connect to that
node and verify that it considers itself master. This is likely to be the
first connection to a node, so any misconfiguration or permission error
will surface here. This commit makes sure to propagate those errors to the
library user.

* Sentinel#replica_for also propagates node coonnection errors.

* DRY logic for which error to return when verifying node role.