Script: Bring script loading into `Script` itself
Loading of the script does not depend on its keys or args (or anything
else from `ScriptInvocation`), so we pull the loading machinery from
`ScriptInvocation` back into `Script`.

Parameter-less scripts allow to get `invoke`d directly on the `Script`
object. But for manually `load`-ing one had to build a
`ScriptInvocation`. Now, we can call `load` directly on the `Script`
object itself.

The `load` and `load_async` methods also stay available on
`ScriptInvocation` to avoid API incompatible changes, and as it is
convenient to have them available, as `ScriptInvocation`'s `script`
field is not public.