Make `Script`'s `load_cmd` public (#1731)
* docs(implement_commands): Drop pipeline reference from `invoke_script`

While `invoke_script` is probably most useful for Pipelines, the
`implement_commands!` macro also implements it for other
`ConnectionLike`s, like `AsyncCommand`.

So we should not refer to `Pipeline`s in its documentation, and we
hence remove this reference.

* ScriptInvocation: Use available helper functions for loading

The `invoke...` methods built the commands to load the script manually,
although there are helper functions for it available.

We switch to these helper functions to bring the different loading
code paths down from 4 to 2. This makes future updates to the script
loading simpler.

* Script: Bring script loading into `Script` itself

Loading of the script does not depend on its keys or args (or anything
else from `ScriptInvocation`), so we pull the loading machinery from
`ScriptInvocation` back into `Script`.

Parameter-less scripts allow to get `invoke`d directly on the `Script`
object. But for manually `load`-ing one had to build a
`ScriptInvocation`. Now, we can call `load` directly on the `Script`
object itself.

The `load` and `load_async` methods also stay available on
`ScriptInvocation` to avoid API incompatible changes, and as it is
convenient to have them available, as `ScriptInvocation`'s `script`
field is not public.

* Pipeline: Add command to load a script

While pipelines could invoke commands, it missed a command to load a
script. So we add one.

---------

Co-authored-by: Christian Aistleitner <christian@quelltextlich.at>