Align async resubscribe behavior.
This changes the behavior of the cluster connection and the connection manager re: resubscribing on subscription that failed.
The connection manager only registers successful subscription and unsubscription - which makes sense. It would be peculiar for the user if a connection that failed subscription would suddenly be subscribed after reconnection to the node.