Merge branch 1.0.x to main (#1796)
* Prepare 1.0.x branch for usage. (#1659)

* Remove unsafe iterators. (#1660)

* Prepare 1.0.x branch for usage.

* t

* Remove unsafe iterators.

* Remove deprecated features. (#1662)

* feat: Make transaction status of pipelines public (#1663)

When trying to mock Redis connections, the `ConnectionLike`'s
`req_packed_commands` expects a different response shape on
`transactional` and non-`transactional` pipelines.

(Vec with single entry of Redis array of results, resp. Vec of results)

In order to hide this difference from users, it would be helpful to
check if a pipeline is in transaction mode or not.

To allow this, we make the accessor function public.

Co-authored-by: Christian Aistleitner <christian@quelltextlich.at>

* Pipeline: Add comments explaining offset and count (#1664)

Co-authored-by: Christian Aistleitner <christian@quelltextlich.at>

* prevent Smol tasks from being cancelled when dropped.

* Remove deprecated features.

---------

Co-authored-by: somechris <somechris@users.noreply.github.com>
Co-authored-by: Christian Aistleitner <christian@quelltextlich.at>

* Add a specific error type for parsing errors. (#1661)

This will save users who just want to parse some data from checking for client or server errors.

* Upgrade criterion version. (#1657)

* Upgrade criterion version.

* increase MSRV

* fix lint

* Remove all deprecated functions. (#1672)

* Make the num-bigint crate optional again. (#1670)

* Make the num-bigint crate optional again.

* Fix no-default-features.

It didn't work because the default features were pulled by redis-test.

* Merge from main (#1680)

* Move TCP features to settings. (#1669)

* Add TCP settings to sync connections.

* Hide internal structure of ConnectionInfo and RedisConnectionInfo.

This will allow us to modify the structs without breaking users code.

* Remove features that are settable through TcpSettings.

* Fix cache index.

The `cache_cmd_index` value was miscalculated, because all tests that checked it had all features enabled, which implicitly removed the CLIENT SETINFO calls form the pipeline.

* fix broken doc links.

* fix cargo config.

* further limit feature combination tests.

* Remove redis setting helper functions.

* fix docs

* Set default timeouts for async connections (#1686)

* Add doc clarifying the addition of timeouts.

* Fix ConnectionManager retry behavior. (#1717)

https://github.com/redis-rs/redis-rs/issues/1556

* Move module files to the `commands` folder. (#1719)

* Move command modules to commands folder.

* Move vector sets structs to separate module.

* Refactor the error types (#1706)

* Move errors to separate modules.

* Clarify type error.

* Fill out  ServerError.

* Use ArcStr in errors, to reduce their size.

* Add ServerError to RedisError variants.

* Add NOPERM error kind.

* Remove repeated variants between ErrorKind and ServerErrorKind.

* Replace some server error kinds with more correct kinds.

* Remove the `ErrorRepr::WithDescription` variant.

It's a repeat of `WithDescriptionAndDetail`.

* Remove `Error` suffix from variants.

* Remove `ErrorRepr::Extension` variant.

It's a repeat of the `ServerError::Extension` variant.

* Ensure internal errors are exposed in `Error::source`.

* Move error constructors to the error modules.

* Allow users to use `ServerError`.

It's already exposed in `Value`, this allows the users to actually use it, too, without exposing its internal implementation.

* Refactor cluster file order and routing. (#1711)

* Move cluster files to shared directory.

* Move shared logic out of sync connection.

* Split slot map from routing.

* Update cluster routing information.

This is based on the work done in the valkey-glide project for mapping commands and their routing info.

* Use builder pattern for SentinelNodeConnectionInfo (#1723)

Align with the changes to `ConnectionInfo`  in https://github.com/redis-rs/redis-rs/pull/1669

* Improve pipeline errors (#1718)

* Move retry_method to server_error.

* Move pipeline response handling to shared function.

* Allow user to receive results or values with server errors per-request.

* Added detailed pipeline error case when user doesn't choose type to RedisResult or Value.

* fix docs

* revert the ability to get all values even on a failure.

* Various minor performance improvements to cluster usage. (#1729)

* Use ArcStr for cluster node addresses.

This reduces the size of allocations, and removes the implied ability to mutate the addresses.

* Improve map usage in clusters.

Removed some repeated key checks.

* Improve slot parsing.

Remove vector splitting and unnecessary ConnectionAddr creation.

* Use ArcStr in ClusterParams, to reduce size.

* Replace some more usage of String with ArcStr.

* Enhance visibility of typed traits in docs. (#1732)

We want more users to be aware of the typed traits, in order to save issues like https://github.com/redis-rs/redis-rs/issues/1730.

* Allow async fire and forget. (#1756)

Allows users to use `set_no_response` in order to skip waiting for the response.
https://github.com/redis-rs/redis-rs/issues/1312

fixes https://github.com/redis-rs/redis-rs/issues/1312

* Add a detailed TransactionAborted error. (#1755)

* Add a detailed TransactionAborted error.

This allows users to see all server errors returned during a transaction, instead of only the first error.

* improve documentation.

* Reduce number of clones. (#1759)

* Reduce number of clones.

Use references instead of some clones.

* Increase MSRV and fix lints.

* Align async resubscribe behavior. (#1757)

* Align async resubscribe behavior.

This changes the behavior of the cluster connection and the connection manager re: resubscribing on subscription that failed.
The connection manager only registers successful subscription and unsubscription - which makes sense. It would be peculiar for the user if a connection that failed subscription would suddenly be subscribed after reconnection to the node.

* fix comments.

* Update version

* fix lints.

* fix: Reject wildcard addresses in client and cluster connections (#1762)

Rejects wildcard addresses like '0.0.0.0' and '::' in connection URLs to prevent connection timeouts, providing a clearer error message. Additionally, when parsing cluster slots, treat wildcard addresses advertised by nodes as the address of the responding node, improving robustness against misconfigured clusters. Fixes #1492

* Create a macro for async tests. (#1779)

This saves a lot of repeated code in test headers.

* Handle more tests in testing macro. (#1783)

* CI: Only run builds after lints pass. (#1785)

This is done in order to reduce the cost of CI on failures.

* Make FromRedisValue take an owned value by default. (#1784)

* Rename functions to reflect priorities.

* Make `from_redis_value` the trait function that users must implement.

This improves the chance of users falling into the pit of success - instead of accidentally triggering allocations, they will reuse the values, if possible.

* add documentation.

* format fixes

* Enable more ways to build a parsing error. (#1788)

This should allow users to build these errors on their implementations of `FromRedisValue`.

* fix version

---------

Co-authored-by: somechris <somechris@users.noreply.github.com>
Co-authored-by: Christian Aistleitner <christian@quelltextlich.at>
Co-authored-by: Cả thế giới là Rust <90588855+naoNao89@users.noreply.github.com>