Support compare and set and compare and delete - SET
Add support for the following value comparisons in the SET command:
IFEQ <match-value> - Set the key's value and expiration only if its current value is equal to <match-value>.
If the key doesn't exist, it won't be created.
IFNE <match-value> - Set the key's value and expiration only if its current value is not equal to <match-value>.
If the key doesn't exist, it will be created.
IFDEQ <match-digest> - Set the key's value and expiration only if the digest of its current value is equal to <match-digest>.
If the key doesn't exist, it won't be created.
IFDNE <match-digest> - Set the key's value and expiration only if the digest of its current value is not equal to <match-digest>.
If the key doesn't exist, it will be created.