Support Compare and set [CAS] and Compare and delete [CAD] (#1862)
* Support compare and set and compare and delete - SET

Add support for the following value comparisons in the SET command:
IFEQ <match-value> - Set the key's value and expiration only if its current value is equal to <match-value>.
If the key doesn't exist, it won't be created.
IFNE <match-value> - Set the key's value and expiration only if its current value is not equal to <match-value>.
If the key doesn't exist, it will be created.
IFDEQ <match-digest> - Set the key's value and expiration only if the digest of its current value is equal to <match-digest>.
If the key doesn't exist, it won't be created.
IFDNE <match-digest> - Set the key's value and expiration only if the digest of its current value is not equal to <match-digest>.
If the key doesn't exist, it will be created.

* Support compare and set and compare and delete - DIGEST

Add the DIGEST command along with its corresponding tests.
Add tests for the SET command with IFDEQ and IFDNE value comparisons.

Provide a way to calculate XXH3 digests by utilizing the xxhash-rust
crate.

* Support compare and set and compare and delete - DELEX

Add the DELEX command.
Add tests that verify the correctness of the DELEX command.

* Add constructors for the ValueComparison enum

Introduce constructors to the ValueComparison,
removing the need for user-managed serialization.

---------

Co-authored-by: StefanPalashev <stefan.palashev@redis.com>